+++ doc/whats-new.rst
@@ -30,7 +30,7 @@ New Features
 - :py:meth:`~xarray.DataArray.rank` now operates on dask-backed arrays, assuming
   the core dim has exactly one chunk. (:pull:`8475`).
   By `Maximilian Roos <https://github.com/max-sixty>`_.
-- Add option ``plot_unit_brackets`` to configure the enclosing brackets of units in plots (:pull:`8503`).
+- Add option ``plot_unit_format`` to configure how units are formatted in plots (:pull:`8503`).
   By `Michael Niklas <https://github.com/headtr1ck>`_.
 
 Breaking changes